Move separate utility method into its own file.
